move exception type spesific error handling from AjaxController to LayerAdminHandler
